fix(web): rename reconnect secret store module so artifact smoke stays clean
The chunk filename reconnect-secret-store-*.js landed in the workbox
precache manifest inside sw.js, where the artifact smoke's
secret-assignment pattern matched the word 'secret' followed by the
next manifest entry's quoted URL. Rename the module (not the check) —
the credential wording also matches the possession-credential design.
